refact: Change JS logic around business types to use the localized data
