+++ client/onboarding-experiment/tasks/add-business-info-task/index.tsx
@@ -10,16 +10,10 @@ import { __ } from '@wordpress/i18n';
 import WizardTaskItem from 'additional-methods-setup/wizard/task-item';
 import WizardTaskContext from 'additional-methods-setup/wizard/task/context';
 import CustomSelectControl from 'components/custom-select-control';
-import { LoadableBlock } from 'components/loadable';
-import { useBusinessTypes } from 'onboarding-experiment/hooks';
 import RequiredVerificationInfo from './required-verification-info';
 import strings from 'onboarding-experiment/strings';
-import {
-	Country,
-	BusinessType,
-	BusinessStructure,
-	OnboardingProps,
-} from 'onboarding-experiment/types';
+import { OnboardingProps } from 'onboarding-experiment/types';
+import { getBusinessTypes } from 'onboarding-experiment/utils';
 
 interface TaskProps {
 	onChange: ( data: Partial< OnboardingProps > ) => void;
@@ -27,7 +21,7 @@ interface TaskProps {
 
 const AddBusinessInfoTask = ( { onChange }: TaskProps ): JSX.Element => {
 	const { isCompleted, setCompleted } = useContext( WizardTaskContext );
-	const { countries, isLoading } = useBusinessTypes();
+	const countries = getBusinessTypes();
 
 	const [ businessCountry, setBusinessCountry ] = useState< Country >();
 	const [ businessType, setBusinessType ] = useState< BusinessType >();
@@ -80,60 +74,54 @@ const AddBusinessInfoTask = ( { onChange }: TaskProps ): JSX.Element => {
 			<p className="complete-business-info-task__subheading">
 				{ strings.onboarding.description }
 			</p>
-			<LoadableBlock isLoading={ isLoading } numLines={ 4 }>
+			<CustomSelectControl
+				label={ __( 'Country', 'woocommerce-payments' ) }
+				value={ businessCountry }
+				onChange={ ( { selectedItem } ) =>
+					handleBusinessCountryUpdate( selectedItem )
+				}
+				options={ countries }
+			/>
+			<p className="complete-business-info-task__description">
+				{ strings.onboarding.countryDescription }
+			</p>
+			{ businessCountry && (
 				<CustomSelectControl
-					label={ __( 'Country', 'woocommerce-payments' ) }
-					value={ businessCountry }
+					label={ __( 'Business type', 'woocommerce-payments' ) }
+					value={ businessType }
+					options={ businessCountry.types }
+					placeholder={ __(
+						'What type of business do you run?',
+						'woocommerce-payments'
+					) }
 					onChange={ ( { selectedItem } ) =>
-						handleBusinessCountryUpdate( selectedItem )
+						handleBusinessTypeUpdate( selectedItem )
 					}
-					options={ countries }
-				/>
-				<p className="complete-business-info-task__description">
-					{ strings.onboarding.countryDescription }
-				</p>
-				{ businessCountry && (
-					<CustomSelectControl
-						label={ __( 'Business type', 'woocommerce-payments' ) }
-						value={ businessType }
-						options={ businessCountry.types }
-						placeholder={ __(
-							'What type of business do you run?',
-							'woocommerce-payments'
-						) }
-						onChange={ ( { selectedItem } ) =>
-							handleBusinessTypeUpdate( selectedItem )
-						}
-					>
-						{ ( item ) => (
-							<div>
-								<div>{ item.name }</div>
-								<div className="complete-business-info-task__option-description">
-									{ item.description }
-								</div>
+				>
+					{ ( item ) => (
+						<div>
+							<div>{ item.name }</div>
+							<div className="complete-business-info-task__option-description">
+								{ item.description }
 							</div>
-						) }
-					</CustomSelectControl>
-				) }
-				{ businessType && businessType.structures?.length > 0 && (
-					<CustomSelectControl
-						label={ __(
-							'Business Structure',
-							'woocommerce-payments'
-						) }
-						value={ businessStructure }
-						options={ businessType.structures }
-						placeholder={ __(
-							'What’s the legal structure of your business?',
-							'woocommerce-payments'
-						) }
-						onChange={ ( { selectedItem } ) =>
-							handleBusinessStructureUpdate( selectedItem )
-						}
-					/>
-				) }
-			</LoadableBlock>
-			<LoadableBlock isLoading={ isLoading } numLines={ 4 } />
+						</div>
+					) }
+				</CustomSelectControl>
+			) }
+			{ businessType && businessType.structures?.length > 0 && (
+				<CustomSelectControl
+					label={ __( 'Business Structure', 'woocommerce-payments' ) }
+					value={ businessStructure }
+					options={ businessType.structures }
+					placeholder={ __(
+						'What’s the legal structure of your business?',
+						'woocommerce-payments'
+					) }
+					onChange={ ( { selectedItem } ) =>
+						handleBusinessStructureUpdate( selectedItem )
+					}
+				/>
+			) }
 			{ businessCountry && businessType && isCompleted && (
 				<RequiredVerificationInfo
 					country={ businessCountry.key }
