+++ client/onboarding-experiment/types.d.ts
@@ -1,21 +1,3 @@
-export interface Country {
-	key: string;
-	name: string;
-	types: BusinessType[];
-}
-
-export interface BusinessType {
-	key: string;
-	name: string;
-	description: string;
-	structures: BusinessStructure[];
-}
-
-export interface BusinessStructure {
-	key: string;
-	name: string;
-}
-
 export interface OnboardingProps {
 	country: string;
 	type: string;
